fix: Resolve all clippy warnings for CI compliance
- Fix format string warnings by using inline formatting
- Replace useless vec! with array syntax
- Fix collapsible if-let patterns
- Replace redundant pattern matching with is_err()
- Fix non-binding let on futures with explicit drop
- Fix drop of references (use let _ instead)
- Add #[allow(dead_code, unused_variables)] to example files
- Add #[allow(dependency_on_unit_never_type_fallback)] for never type warning
- Remove unnecessary enumerate() when index is unused

All clippy checks now pass with --all-targets --all-features.

+++ crates/rustyray-core/tests/actor_task_integration.rs
@@ -326,7 +326,7 @@ async fn test_concurrent_actors_and_tasks() {
         // Each actor with id i creates values [i, 2*i, 3*i, 4*i, 5*i]
         // So sum should be i * (1+2+3+4+5) = i * 15
         for (idx, sum) in results {
-            assert_eq!(sum, (idx as i32) * 15);
+            assert_eq!(sum, idx * 15);
         }
     })
     .await;
